api/pyproject.toml
[project]
name = "api"
version = "0.2.0"
description = "Exam System API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.2",
    "faker>=40.12.0",
    "anthropic>=0.88.0",
    "pytesseract>=0.3.13",
    "pillow>=12.2.0",
    "locust>=2.43.4",
    "orjson>=3.11.8",
]
